feat(language-picker): generate correct links for blog posts
All post markdown files now contain a altLocaleUrl
frontmatter atttribute that specifies their alternate
locale href path.
The language-picker template now checks for the
existance of altLocaleUrl and sets it as the
href if it exists.
Otherwise, it passes page.url to the swap_locale
Nunjucks filter, which has replaced the locale_url
filter.